# [sdk] Perf improvement for TraceContextPropagator.Inject