## Changes

* Use [string.Create](https://learn.microsoft.com/dotnet/api/system.string.create#system-string-create-1(system-int32-0-system-buffers-spanaction((system-char-0)))) (where available) when building `traceparent` to improve the performance of `TraceContextPropagator.Inject`

## Benchmarks

I tested a bunch of different ways we could do it:

```csharp
[MemoryDiagnoser]
public class TraceParentBenchmarks
{
    private static readonly ActivityContext Context = new(ActivityTraceId.CreateRandom(), ActivitySpanId.CreateRandom(), ActivityTraceFlags.Recorded);

    [Benchmark]
    public string BuildUsingStringConcat()
    {
        var traceparent = string.Concat("00-", Context.TraceId.ToHexString(), "-", Context.SpanId.ToHexString());
        traceparent = string.Concat(traceparent, (Context.TraceFlags & ActivityTraceFlags.Recorded) != 0 ? "-01" : "-00");
        return traceparent;
    }

    [Benchmark]
    public string BuildUsingStringInterpolation()
    {
        return $"00-{Context.TraceId.ToHexString()}-{Context.SpanId.ToHexString()}-{((Context.TraceFlags & ActivityTraceFlags.Recorded) != 0 ? "-01" : "-00")}";
    }

#if !NETFRAMEWORK
    [Benchmark]
    public string BuildUsingStringCreateAndInterpolation()
    {
        return string.Create(
            null,
            stackalloc char[128],
            $"00-{Context.TraceId.ToHexString()}-{Context.SpanId.ToHexString()}-{((Context.TraceFlags & ActivityTraceFlags.Recorded) != 0 ? "-01" : "-00")}");
    }

    [Benchmark]
    public string BuildUsingStringCreate()
    {
        return string.Create(55, Context, BuildTraceParent);

        static void BuildTraceParent(Span<char> destination, ActivityContext context)
        {
            "00-".CopyTo(destination);
            context.TraceId.ToHexString().CopyTo(destination.Slice(3));
            destination[35] = '-';
            context.SpanId.ToHexString().CopyTo(destination.Slice(36));
            if ((context.TraceFlags & ActivityTraceFlags.Recorded) != 0)
            {
                "-01".CopyTo(destination.Slice(52));
            }
            else
            {
                "-00".CopyTo(destination.Slice(52));
            }
        }
    }

    [Benchmark]
    public string BuildUsingStringCtor()
    {
        Span<char> destination = stackalloc char[55];

        "00-".CopyTo(destination);
        Context.TraceId.ToHexString().CopyTo(destination.Slice(3));
        destination[35] = '-';
        Context.SpanId.ToHexString().CopyTo(destination.Slice(36));
        if ((Context.TraceFlags & ActivityTraceFlags.Recorded) != 0)
        {
            "-01".CopyTo(destination.Slice(52));
        }
        else
        {
            "-00".CopyTo(destination.Slice(52));
        }

        return new(destination);
    }
#endif
}
```

This is on .NET 8:

| Method                                 | Mean     | Error    | StdDev   | Gen0   | Allocated |
|--------------------------------------- |---------:|---------:|---------:|-------:|----------:|
| BuildUsingStringConcat (existing method)   | 19.25 ns | 0.253 ns | 0.236 ns | 0.0210 |     264 B |
| BuildUsingStringInterpolation          | 30.03 ns | 0.351 ns | 0.328 ns | 0.0108 |     136 B |
| BuildUsingStringCreateAndInterpolation | 21.66 ns | 0.462 ns | 0.409 ns | 0.0108 |     136 B |
| BuildUsingStringCreate (this pr)      | **13.83 ns** | 0.197 ns | 0.174 ns | 0.0108 |     **136 B** |
| BuildUsingStringCtor                   | 15.24 ns | 0.152 ns | 0.142 ns | 0.0108 |     136 B |

## Merge requirement checklist

* [X] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-dotnet/blob/main/CONTRIBUTING.md) guidelines followed (license requirements, nullable enabled, static analysis, etc.)